requirements.txt
pyyaml
ansible
pywinrm
